requirements.txt
python-pptx==1.0.2
loguru==0.7.2
gradio==4.43.0
markdown==3.7
langchain==0.2.16
langchain_core==0.2.41
langchain_community==0.2.17
langchain_openai==0.1.25
loguru==0.7.2